+++ backend/src/validations/user.ts
@@ -4,16 +4,17 @@ import { body } from "express-validator";
 export const signupValidation = [
   body("fullName").notEmpty().withMessage("Name is required"),
   body("email").isEmail().withMessage("Invalid email"),
-  body("password").isLength({ min: 6 }).withMessage("Password must be at least 6 characters"),
-  // TODO: uncheck in production
-  // .matches(/\d/)
-  // .withMessage("Password must contain at least one number")
-  // .matches(/[A-Z]/)
-  // .withMessage("Password must contain at least one uppercase letter")
-  // .matches(/[a-z]/)
-  // .withMessage("Password must contain at least one lowercase letter")
-  // .matches(/[!@#$%^&*(),.?":{}|<>]/)
-  // .withMessage("Password must contain at least one special character"),
+  body("password")
+    .isLength({ min: 6 })
+    .withMessage("Password must be at least 6 characters")
+    .matches(/\d/)
+    .withMessage("Password must contain at least one number")
+    .matches(/[A-Z]/)
+    .withMessage("Password must contain at least one uppercase letter")
+    .matches(/[a-z]/)
+    .withMessage("Password must contain at least one lowercase letter")
+    .matches(/[!@#$%^&*(),.?":{}|<>]/)
+    .withMessage("Password must contain at least one special character"),
 ];
 
 export const loginValidation = [