+++ client/src/pages/Dashboard.tsx
@@ -13,7 +13,17 @@ import HomeWorkspaceCard from "@/layouts/dashboard/HomeWorkspaceCard";
 import { FormType, WorkspaceType } from "@/lib/types";
 import { fetchDashboard } from "@/services/dashboard";
 import { useQuery } from "@tanstack/react-query";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
+import {
+  Empty,
+  EmptyContent,
+  EmptyDescription,
+  EmptyHeader,
+  EmptyMedia,
+  EmptyTitle,
+} from "@/components/ui/empty";
+import { FileText } from "lucide-react";
+import { Button } from "@/components/ui/button";
 
 const DashboardPage = () => {
   const {
@@ -25,36 +35,96 @@ const DashboardPage = () => {
     queryFn: fetchDashboard,
     staleTime: 10000,
   });
-  // const publishedForms: FormType[] = dashboard?.workspaces?.flatMap((workspace: WorkspaceType) =>
-  //   workspace.forms
-  //     .filter(form => form.published)
-  //     .map(form => ({ ...form, workspaceId: workspace._id })),
-  // );
+  const navigate = useNavigate();
+
   const allForms: FormType[] = dashboard?.workspaces?.flatMap((workspace: WorkspaceType) =>
     workspace.forms.map(form => ({ ...form, workspaceId: workspace._id })),
   );
-  // const draftedForms: FormType[] = dashboard?.workspaces?.flatMap(
-  //   (workspace: WorkspaceType) =>
-  //     workspace.forms
-  //       .filter(form => !form.published)
-  //       .map(form => ({ ...form, workspaceId: workspace._id })), // Only include forms that are not published
-  // );
+
+  const workspaces: WorkspaceType[] = dashboard?.workspaces || [];
+
+  const handleClick = () => {
+    navigate(`/form?workspaceId=${workspaces[0]._id}`); // Replace with your desired route
+  };
   return (
     <div className="px-5">
       <PageTitle>Home</PageTitle>
       <main className="mx-auto max-w-[1100px] overflow-auto flex-grow container ">
         {!isError && (
-          <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
-            <HomeWorkspaceCard
-              showSkeleton={!isError && isLoading}
-              workspaceCount={dashboard?.workspaces?.length}
-            />
-            <HomeStatsCard
-              isLoading={isLoading}
-              workspaces={dashboard?.workspaces}
-            />
+          <div className="">
+            <h1 className="text-lg font-semibold">Quickstart</h1>
+            <div className="flex gap-2 mt-4">
+              <HomeWorkspaceCard />
+              <HomeStatsCard isLoading={isLoading} />
+            </div>
           </div>
         )}
+        <Tabs
+          defaultValue="all"
+          className="mt-12"
+        >
+          <TabsList>
+            <TabsTrigger value="all">Workspaces</TabsTrigger>
+          </TabsList>
+          <TabsContent
+            value="all"
+            className="mt-4"
+          >
+            <div className="grid gap-4">
+              {!isError && dashboard?.workspaces?.length > 0 ? (
+                <Table>
+                  <TableHeader>
+                    <TableRow>
+                      <TableHead className="w-[20%]">Name</TableHead>
+                      <TableHead className="w-[20%]">Forms</TableHead>
+                      <TableHead className="w-[20%]">Created</TableHead>
+                      <TableHead className="w-[20%] text-right">Actions</TableHead>
+                    </TableRow>
+                  </TableHeader>
+                  <TableBody>
+                    {workspaces.map(workspace => (
+                      <TableRow
+                        key={workspace._id}
+                        className="group"
+                      >
+                        <TableCell className="w-[20%] font-bold">
+                          <Link
+                            to={`/workspaces/${workspace._id}?name=${workspace.name}`}
+                            onClick={e => e.stopPropagation()}
+                            className="group-hover:underline"
+                          >
+                            {workspace.name}
+                          </Link>
+                        </TableCell>
+                        {/* <TableCell className="w-[20%]">
+                          {form.published ? "Published" : "Draft"}
+                        </TableCell> */}
+                        <TableCell className="w-[20%]">{workspace.forms.length}</TableCell>
+                        <TableCell className="w-[20%]">
+                          {new Date(workspace.created).toLocaleDateString("en-US")}
+                        </TableCell>
+                        <TableCell className="w-[20%] text-right space-x-3 font-semibold text-primary">
+                          <Link
+                            to={`/form?workspaceId=${workspace._id}`}
+                            onClick={e => e.stopPropagation()}
+                            className="hover:underline"
+                          >
+                            Create Form
+                          </Link>
+                        </TableCell>
+                      </TableRow>
+                    ))}
+                  </TableBody>
+                </Table>
+              ) : (
+                <p className="text-2xl mx-auto w-fit my-16 font-semibold text-muted-foreground">
+                  No workspaces available
+                </p>
+              )}
+            </div>
+          </TabsContent>
+        </Tabs>
+        {/* Forms overview table */}
         <Tabs
           defaultValue="all"
           className="mt-12"
@@ -84,7 +154,6 @@ const DashboardPage = () => {
                   <TableHeader>
                     <TableRow>
                       <TableHead className="w-[20%]">Name</TableHead>
-                      {/* <TableHead className="w-[20%]">Status</TableHead> */}
                       <TableHead className="w-[20%]">Submissions</TableHead>
                       <TableHead className="w-[20%]">Created</TableHead>
                       <TableHead className="w-[20%] text-right">Actions</TableHead>
@@ -105,9 +174,6 @@ const DashboardPage = () => {
                             {form.name}
                           </Link>
                         </TableCell>
-                        {/* <TableCell className="w-[20%]">
-                          {form.published ? "Published" : "Draft"}
-                        </TableCell> */}
                         <TableCell className="w-[20%]">{form.submissions}</TableCell>
                         <TableCell className="w-[20%]">
                           {new Date(form.created).toLocaleDateString("en-US")}
@@ -135,98 +201,33 @@ const DashboardPage = () => {
                   </TableBody>
                 </Table>
               ) : (
-                <p className="text-2xl mx-auto w-fit my-16 font-semibold text-muted-foreground">
-                  No published forms available
-                </p>
+                <Empty>
+                  <EmptyHeader>
+                    <EmptyMedia variant="icon">
+                      <FileText />
+                    </EmptyMedia>
+                  </EmptyHeader>
+                  <EmptyTitle>No Forms Yet</EmptyTitle>
+                  <EmptyDescription>
+                    You haven&apos;t created any forms yet. Get started by creating your first form.
+                  </EmptyDescription>
+                  <EmptyContent>
+                    <div className="flex gap-2">
+                      <Button onClick={handleClick}>Create Form</Button>
+                    </div>
+                  </EmptyContent>
+                </Empty>
               )}
             </div>
           </TabsContent>
           <TabsContent
             value="published"
             className="mt-4"
-          >
-            {/* <div className="grid gap-4">
-              {!isError && publishedForms?.length > 0 ? (
-                publishedForms.map((form) => (
-                  <Card key={form._id}>
-                    <CardHeader>
-                      <CardTitle>{form.name}</CardTitle>
-                      <CardDescription>
-                        Status: <span className="capitalize">published</span>
-                      </CardDescription>
-                    </CardHeader>
-                    <CardContent>
-                      <p>Submissions: {form.submissions}</p>
-                    </CardContent>
-                    <CardFooter className="flex justify-between">
-                      <Button variant="outline">
-                        <FileText className="mr-2 h-4 w-4" /> View
-                      </Button>
-                      <div className="flex align-middle gap-5">
-                        {form.published && (
-                          <Button variant={"ghost"}>
-                            <LinkIcon />
-                          </Button>
-                        )}
-                        <Link to={`/form/${form.form_id}`}>
-                          <Button variant="outline">
-                            <Settings className="mr-2 h-4 w-4" /> Edit
-                          </Button>
-                        </Link>
-                      </div>
-                    </CardFooter>
-                  </Card>
-                ))
-              ) : (
-                <p className="text-2xl mx-auto w-fit mt-28 font-semibold mb-2 text-muted-foreground">
-                  No published forms available
-                </p>
-              )}
-            </div> */}
-          </TabsContent>
+          />
           <TabsContent
             value="drafts"
             className="mt-4"
-          >
-            {/* <div className="grid gap-4">
-              {!isError && draftedForms?.length > 0 ? (
-                draftedForms.map((form) => (
-                  <Card key={form._id}>
-                    <CardHeader>
-                      <CardTitle>{form.name}</CardTitle>
-                      <CardDescription>
-                        Status: <span className="capitalize">drafted</span>
-                      </CardDescription>
-                    </CardHeader>
-                    <CardContent>
-                      <p>Submissions: {form.submissions}</p>
-                    </CardContent>
-                    <CardFooter className="flex justify-between">
-                      <Button variant="outline">
-                        <FileText className="mr-2 h-4 w-4" /> View
-                      </Button>
-                      <div className="flex align-middle gap-5">
-                        {form.published && (
-                          <Button variant={"ghost"}>
-                            <LinkIcon />
-                          </Button>
-                        )}
-                        <Link to={`/form/${form.form_id}`}>
-                          <Button variant="outline">
-                            <Settings className="mr-2 h-4 w-4" /> Edit
-                          </Button>
-                        </Link>
-                      </div>
-                    </CardFooter>
-                  </Card>
-                ))
-              ) : (
-                <p className="text-2xl mx-auto w-fit mt-28 font-semibold mb-2 text-muted-foreground">
-                  No published forms available
-                </p>
-              )}
-            </div> */}
-          </TabsContent>
+          />
         </Tabs>
       </main>
     </div>