+++ backend/CHANGELOG.md
@@ -1,5 +1,41 @@
 # backend
 
+## 0.12.0
+
+### Minor Changes
+
+- 0fd9e44: changelogs in production
+- 0fd9e44: Stable beta version ready for prod
+- 0fd9e44: production ready tailwind fix
+
+## 0.11.0
+
+### Minor Changes
+
+- ddc2a32: changelogs in production
+- ddc2a32: Stable beta version ready for prod
+- ddc2a32: production ready tailwind fix
+
+## 0.10.0
+
+### Minor Changes
+
+- 8499184: changelogs in production
+- 8499184: Stable beta version ready for prod
+- 8499184: production ready tailwind fix
+
+## 0.9.0
+
+### Minor Changes
+
+- a51d066: changelogs in production
+
+## 0.8.0
+
+### Minor Changes
+
+- 605c8a8: production ready tailwind fix
+
 ## 0.7.0
 
 ### Minor Changes